refactor(admin): fold storage backend into admin
This makes it clearer which data is stored where.
