+++ frontends/election-manager/src/lib/backends/index.ts
@@ -1,5 +1,4 @@
 /* istanbul ignore file */
 export * from './admin_backend';
 export * from './memory_backend';
-export * from './storage_backend';
 export * from './types';
